feat: implement ExecutorChain
Implement ExecutorChain class which contains a collection of nested Executor classes (which extend the AbstractExecutor).
The ExecutorChain instantiates all Executor classes and assembles a chain of responsibility.
Each Executor class corresponds to a valid opcode from the Bytecodes class.
Each Executor class is responsible for the execution of a specific opcode. All executor classes interract with the stack and instruction pointers of the VM class.
List of partially implemented Executor classes: IADD, IMUL, IDIV, ISUB, LREF, ICONST, LLOAD, LSTORE, HALT. The rest are not yet implemented.